feat: Refactor main.js and scripts for improved structure; add initMenuAnimation, mountSprinkledVueApps, and mountTimelineApp functions
