+++ src/scripts/app.js
@@ -1,7 +1,5 @@
 import { createApp } from 'vue';
-import 'stickUp';
 import IssueTimeline from './vue/IssueTimeline.vue';
-import _ from 'underscore';
 
 
 
@@ -11,14 +9,6 @@ import _ from 'underscore';
  * @type {boolean}
  */
 const app = createApp(IssueTimeline);
-
-
-
-/**
- * StickUp
- */
-
-$('.navbar').stickUp();
 
 /**
  * SearchBar
@@ -40,6 +30,9 @@ window.searchToggle = searchToggle;
  */
 
 document.addEventListener('DOMContentLoaded', function () {
+
+
+    console.log('Anime App Loaded');
 
     app.mount('#timeline-App');
 
@@ -96,29 +89,3 @@ function closest(el, selector) {
 
     return null;
 }
-
-document.addEventListener('DOMContentLoaded', function () {
-    // Get all document sliders
-    var showQuickview = document.querySelectorAll('[data-show="quickview"]');
-    [].forEach.call(showQuickview, function (show) {
-        var quickview = document.getElementById(show.dataset['target']);
-        if (quickview) {
-            // Add event listener to update output when slider value change
-            show.addEventListener('click', function (event) {
-                quickview.classList.add('is-active');
-            });
-        }
-    });
-
-    // Get all document sliders
-    var dismissQuickView = document.querySelectorAll('[data-dismiss="quickview"]');
-    [].forEach.call(dismissQuickView, function (dismiss) {
-        var quickview = closest(dismiss, '.quickview');
-        if (quickview) {
-            // Add event listener to update output when slider value change
-            dismiss.addEventListener('click', function (event) {
-                quickview.classList.remove('is-active');
-            });
-        }
-    });
-});
